requirements.txt
flask
requests
python-dotenv
gunicorn
flask-cors
discord.py
aiohttp